requirements.txt
absl-py==2.1.0
accelerate==0.27.2
addict==2.4.0
aiofiles==23.2.1
altair==5.2.0
anyio==4.3.0
attrs==23.2.0
certifi
cffi
charset-normalizer
click==8.1.7
colorama==0.4.6
contourpy==1.1.0
cryptography
cycler==0.11.0
deepspeed==0.14.0
einops==0.7.0
fastapi==0.110.0
ffmpy==0.3.2
filelock
fonttools==4.42.1
fsspec==2023.6.0
gradio==4.21.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
huggingface-hub==0.21.4
idna
imageio==2.34.0
importlib-metadata==6.8.0
ipykernel
ipython
ipywidgets==8.1.2
jedi
Jinja2
joblib==1.3.2
jsonschema==4.21.1
jupyter-client
jupyter_core
kiwisolver==1.4.4
lap
lit==15.0.7
Markdown==3.5.2
MarkupSafe
matplotlib==3.7.2
mmcls==1.0.0rc6
mmcv
mmdet==3.0.0rc0
mmengine==0.4.0
motmetrics==1.4.0
moviepy==1.0.3
networkx
numpy
openai==1.14.1
opencv-python==4.8.0.76
orjson==3.9.15
packaging==23.1
pandas==1.3.5
Pillow==9.4.0
protobuf==5.26.0
psutil
pydantic==2.6.4
pydub==0.25.1
pynvml==11.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
regex==2023.8.8
requests
rich==13.7.1
scikit-learn==1.4.1.post1
scipy==1.7.3
seaborn==0.13.2
six
starlette==0.36.3
sympy
tabulate==0.9.0
tensorboard==2.16.2
termcolor==2.4.0
thop==0.1.1.post2209072238
threadpoolctl==3.3.0
tokenizers==0.13.3

tqdm==4.66.1
transformers==4.31.0
typer==0.9.0
typing_extensions==4.10.0
ultralytics==8.1.29
urllib3
uvicorn==0.28.0
Werkzeug==3.0.1
xmltodict==0.13.0